Решили вопрос с переполнением поменяв int на double Р# Please enter the commit message for your changes. Lines starting

+++ Ex043_rekursia/Program.cs
@@ -1,4 +1,4 @@
-﻿int Factorial(int n)
+﻿double Factorial(int n)
 {
     // 1! = 1
     // 0! = 1
